proj1/src/main/java/agh/ii/prinjava/proj1/impl/MyQueueDLLBImpl.java
package agh.ii.prinjava.proj1.impl;

import agh.ii.prinjava.proj1.MyQueue;

/**
 * The goal of this class is to implement a queue
 * @param <E>, the elements that we store in the queue can be of different types, for example : int, String ect...
 */
public class MyQueueDLLBImpl<E> implements MyQueue<E> {
    private DLinkList<E> elems;

    /** Instanciates a Double linked list as a new Queue.
     * This method is a constructor of the MyQueueDLLBImpl class.
     */
    public MyQueueDLLBImpl() {
        this.elems = new DLinkList<>();
    }

    /** This method allows to add a new node to final position of queue.
     * It uses the addLast method from DLinkList.java class.
     * @param x : Node to add to last position of queue.
     */
    @Override
    public void enqueue(E x) {
        elems.addLast(x);
    }

    /** This method allows to remove node in first position of queue.
     * It uses the removefirst method from DLinkList.java class.
     * @return the value of the removed node from list
     */
    @Override
    public E dequeue() {
        return elems.removeFirst();
    }

    /** Gets the value of a list's length and returns it.
     * @return integer corresponding to the list's size.
     */
    @Override
    public int numOfElems() {
        return elems.size();
    }

    /** Gets element in first position of queue.
     * @return the value of element on indicated position
     */
    @Override
    public E peek() {
        return elems.peek();
    }

}